Create 5. Conditional statements 2
For each integer n in the inclusive interval [a,b]:

If  1<=n<=9, then print the English representation of it in lowercase. That is "one" for 1 , "two" for 2 , and so on.
Else if  n and it is an even number, then print "even".
Else if  n and it is an odd number, then print "odd".